refactor eligible methods types and move them to their own file
